moved projects as a seperate component for each
